Adds feature flag functionality to the service
As we need to be able to deploy, with the ability to hide certain
features, we have decided to use feature flags rather than juggling
branches to handle this.

The [flipflop
gem](https://github.com/voormedia/flipflop/blob/master/README.md) allows
us to easily add new features (to config/features.rb) and have them
available to turning features on or off.  This is normally done at the
/flipflop URL (where the UI is mounted as a rails engine) but this is
currently disabled until we have a more appropriate role to manage
access.

This PR implements the feature flag for showing the link to
allocation-info on the allocated offenders tab in the summary.