Enables bulk import of case_information
As we onboard new prisons, we end up with a situation where they can
have hundreds and hundreds of offenders whose tier, service provider and
omic eligibility we don't know.  Although we implemented a stop-gap
until data was available from NDelius, we suspect the amount of work
needed for each prison was off-putting.

As each prison currently implementing OMIC uses spreadsheets to maintain
their allocations, we have implemented this script to import a subset of
that data directly into the CaseInformation table.

This is currently an internal/manual task that is an aid to onboarding
prisons until we have access to the delius feed.

The format of the CSV should be

```
ID,Tier,Provider,OMIC
```

although this will mean massaging the prison's files to be in that
format, it is likely to be a lot less work than the task being done
(again).

The current behaviour of the importer is:
    * Complain with line number if any field is missing
    * Refuse to overwrite any existing record for that offender

You can run the last, passing the prison code and the path to the csv
file with

```
rails bulk_upload:import[AAA,test.csv]
```