[POM 86] Use updated_at instead of created_at
When we are determining when a POM was allocated, we will have filtered
to show allocations where that POM is currently the primary.  We then
need to know WHEN they were allocated, and for now we are going to use
the updated_at field, instead of the created_at field.

We will need to look at this again for POM-87 as co-working will
complicate the calculation and require navigating previous versions.