Change how we send any reallocation emails
Currently the way we determine whether to send a reallocation email,
along with the allocation email, results in us always sending a
reallocation email along with the allocation email.  This is because it
just pulls the last allocation which worked before we changed to using
papertrail.

This PR changes how it is sent, and contains slight changes to the
interface.  It is expected that only a message and the recently created
allocation object will be passed to the EmailService.instance() class
method, the other data is derived for this.

To determine if the service should send a reallocated email, we now
check that the newly created allocation is a :reallocated_primary_pom
event, and if so try and determine the previous pom's details.  We do
this by iterating through the previous versions to find the
first primary_pom that is not the pom in the current allocation.

This currently relies on the :reallocated_primary_pom event being
accurate (as it is over-writing a previously allocated POM) when
triggered by the user interface.  This _should_ be true. We should
ensure that there are no further edge cases.

Further refactoing of the EmailService when we look again at all of the
notifications we might send, may want to break this service up to handle
specific events explicitly rather than hiding logic within the service
itself.