Fix integration test cleanup task
We have implemented a new process regarding the way we track
allocations & reallocations.  This has resulted in a new column being
added in the database (AllocationVersions), and this is used instead of
the original Allocations column.

As part of our staging environment we have set up integration tests to
ensure that our workflows are behaving as expected. We then have a
corresponding cron job that runs every night to remove any allocations
created via the integration tests to ensure we don't get to a point
where everyone is allocated and the tests fail.  However, we have now
reached this point because the cron job still refers to the old column,
and not the new one.  This PR therefore updates the cron job to use the
new column name.