Fix seeds file
There have been changes to the AllocationService and additional fields
are now required to be present when creating a new allocation.  This
missing/incorrect information was causing the db:seeds command to fail.
In addition to this it was found that the section where the allocations
were being created wasn't working properly, so allocations weren't
showing up in the correct tabs. Finally
there was an issue with emails being sent due to a malformed email
address in T3 for Ross (used for testing) and so Kath's POM details were
added and these cases are now allocated to her instead.  Following the changes
to the seed file users are able to run the db:seeds command successfully
again.