Destroy CaseInformation records during cleanup
In addition to cleaning up the AllocationVersion records created during
the integration tests we also need to destroy any CaseInformation
records created too.  At present this isn't done via any rake tasks/cron
jobs and every so often we run into a problem where we have
CaseInformation for all offenders and the integration tests fail.  To
date we have been manually deleting these, but it makes sense to include
this in the current cleanup job.