Create rake task for onboarding prisons
In an effort to make the process of onboarding a prison more efficient
we are going to create CaseInformation records from the Delius data
extract.

This commit starts the skeleton of the rake task and fetches all the
offenders for a prison.

It also includes the sample Delius xlxs file we've created to test
against from T3 data.