Final stage ETL, write CaseInformation
As the final stage of the ETL it is necessary to create CaseInformation
objects in the database based on the remaining offenders (that do not
already have records) and populated with the in-memory delius-data.  In
order to make the delius-data lookup fast we convert the list of hashes
into a hash where the key is the nomis id.

The rake task will now perform the final step of the process and report
back on the number of additions, and the number of offenders in the
prison whose nomis_id could not be found in the delius data.