Adds the country where the prison is located
When adding missing information, it is necessary to determine whether
the current offender falls under the OMIC policy.  In Wales this is
determined by the offender's last known LDU being in Wales, but for
English prisons the default is True.

To allow us to onboard English prisons, we will need to hide the
omicable field when adding missing information, or using delius_etl, and
default to True.  This PR makes it possible to determine, by prison
code, whether a Prison is in England or Wales (:england, or :wales).