Loads data from CSV rather than XLSX
The previous attempt at extracting the delius data from an XLSX file,
whilst functional, used excessive amounts of RAM which exceeded that
available in the production environment. This was the case with 2 of the
most popular libraries for reading XLSX in Ruby, most likely as a direct
result of parsing 10s of Mb of XML.

This PR instead switches to reading the delius data from a CSV file with
the same structure.  The results are much faster execution, and lower
memory requirements, although still high because of the need to include
the Rails environment in the ETL task.

In addition to this change, we're now also reducing the data we hold in
the delius lookup structure and we might find further savings by
inverting the processing, so iterating through every line of the CSV and
instead looking up whether it matches an offender.