Move offender allocation history to a new page
Following design changes the allocation history page is to be
moved, so that it will be the only data on the page.

A new route has been created that will display the allocation
history for a specific offender. The view for this route will
be changed but for now it contains the view from:
app/views/shared/_allocation_history.html.erb