Fix case responsibility calculation bug
We have been getting Sentry alerts for an issue around the various POM
caseload views: https://sentry.service.dsd.io/mojds/live1-allocation-manager-produ/issues/36091/.
Investigation into this issue showed that this error appears when
someone has been released and our MovementService hasn’t run yet.
On release the sentence_start_date is set to nil, which is why we’re getting the error.
From the looks of things the error will trigger when
sentence_start_date == nil && release_date == today).

This method is part of the flow where we determine whether the POM is
responsible or not.  We don’t keep a record of this responsibility/supporting
outcome anywhere as we run the checks each time we load the POM caseload page
or the unallocated summary tab, so we can’t rely on grabbing a field out of the db.

At the point we check if someone is sentenced or not we are going to
return false if they 'sentence_start_date' is blank.