Provides an API for fetching a single offender's category
As the single offender endpoint does not contain the category code, and
we are informed that this isn't the place for it, this PR adds a new API
call which can be used as necessary to find the category_code of a
single offender.

This new API call is not currently integrated into the get_offender
call, instead we should use as necessary until such a point as it
becomes obvious it should be part of get_offender.