Do not format the Faker generated names
We were formatting the Faker names using the titlecase keyword, but
Faker names are already in the correct format. This caused failures
due to discrepancies in the formatting of the text expected. For
example 'Peter Pan IV' became 'Peter Pan Iv'.

To fix this issue we are no longer using titlecase, and as such we
no longer need to concatenating the Faker first name and last name
in the tests.