Do not log an error if keyworker fails
If the keyworker API fails, typically with a 404 we currently show Not
Assigned in the views.  Instead we will not log the error, it's a
non-critical API.  Rather than show as Not Assigned, we also change the
text to 'Data not available' to cover cases where the API is actually
broken.

We need to clear the text with LM