Refactors how we fetch bulk sentence details.
Inspired by finding multiple offenders with the same booking ID in T3,
which induced a crash.

To find the sentence details for a group of offenders, using the Elite2
API, we need to post a list of `booking ids`, which is the current
booking id for a given offender.  Once we had fetching this data, we
were then creating a hash where the services could lookup the sentence
details using the nomis_id.  This was slightly confusing, and introduced
a problem when we had to offenders looking up what _should_ have been
the same booking id.

In order to make this API a little more consistent, after passing in
booking IDs to the function the returned hash is also keyed on those
same IDs, rather than the nomis id for the offender.  This means we can
handle the possible situation where two offenders share a booking id
(which will never happen in prod, right?).