Update Allocation Info page
Further design & content changes have been made to the Allocation
Information page.  There continue to a few sections that we haven't
built the functionality for, or able to get the data for; however the
decision has been made to comment them out to ensure we remember to add
this information as soon as we have it.

There is also a feature flag on the allocation history section as we
only want this displayed once the whole feature has been completed.