Adds a blank help page and a blank contact page
We are going to be adding a contact form, and a help page (with static
content) so this PR simply sets up the routes and templates for these
pages.

The contact page will require a form to post (and then email a message)
but the templates are currently there for content.