Change AllocationVersion validated fields
There is no need for the fields `primary_pom_nomis_id` and
`primary_pom_allocated_at` to have to be validated when a
AllocationVersion record is created or updated as these
values can be nil. An example of this is when an offender is
deallocated, they will no longer be allocated to a primary
pom so these fields will be set to nil.