Fix offender deallocation
When an offender moves to a new prison, we want to deallocate them from
their current POM.  This was using update_all to update some fields of
the current allocation for that offender, unfortunately this does not
work with papertrail.  The end result of this not working was that the
current allocation was overwritten, but no new version record appeared.
This breaks the history.

Rather than use update_all, this PR switches to setting the specific
fields in the allocation_version and then re-saving it.