Fix names with apostrophes cause test failures
Sometimes the tests fail due to the expectation of a name, where
that name has an apostrophe, for example "Bob O'Shea" will
appear as "Bob O'shea". This is not correct the name should be
"Bob O'Shea".

Using the methods titleize or titlecase causes the letter after
the apostrophe to be lowercase, which will cause the expectation
failure.

For consistency we no longer use the titlecase method as it
pretty much does the same thing as titleize.

To solve this issue the titleize method has been overwritten so
that it does not lower case the letter displayed after the
apostrophe.

The bonus of using this overridden method is that previously a
double barrelled name would remove the hyphen which is not ideal.
For example "Bob Cullen-Smith" would have been converted to
"Bob Cullen Smith". Now the hyphen remains in place which is
more accurate.