Implement new parser for delius continuous import
For the continuous import we need to be able to parse data out of an
XLSX file, but without using more memory than is available in the
runtime pod.  Ideally we'd also run without the conversion to CSV as
this might also be problematic (as it still involves parsing the XLSX
file).

This PR implements a [SAX
parser](https://en.wikipedia.org/wiki/Simple_API_for_XML) which can
parse the data out of the XLSX using a lot less memory than those
solutions that use a DOM parser.  XLSX files are basically ZIP files
that contain XML files, but DOM parsing requires that the entire file be
loaded in memory (or represented in memory) which uses a lot of RAM when
the XML files are expanded from the ZIP.

Unfortunately the main sheets only index into a sharedStrings.xml file
so we have to load all of the data from that (again using SAX) before we
can return any data to the caller.

Still, this approach has many benefits:

* 300% faster than previously
* 40% of the RAM usage
* Same block/callback based approach

Although this PR does not currently write to the database, the hooks are
all in place to enable it in a new rake task called
delius_import:load which will populate the table once it exists.