Adds a table for inserting current delius data.
During the continuous update of offender case information, we will be
importing delius data for fast lookup/setting of the case information.

As we do not wish to query for each row, then determine what has changed
and do an insert or update as appropriate, this PR takes more direct
approach of using INSERT ON CONFLICT to insert the data and update just
those fields that we care about (currently tier).  If the tier during an
update changes from the previous value, a record is written into the
tier_changes table (via the TierChange model). There is no restriction
on the number of Tier changes for an offender, and as they are tagged
with an updated_at field, this can be used for the history.