Switch feature flag on for prisoner profile & allocation history pages
We have implemented the designs for the prisoner profile & allocation
history pages.  As these are accessed via a change in the workflow they
have been hidden behind a feature flag.  We now want to start displaying
these pages so that we can get sign off/testing from the team, before we
push these out to production.  This PR therefore switches on the feature
flag and amends the relevant feature spec to include these new pages.