Fix deallocate primary POM
When a POM leaves, or is deactivated, we want to remove their details
from all their current allocations.  To date we have been using
update_all to update some of the fields in the current allocation for
all offenders.  However, this doesn't work with the papertrail gem and
in it's current state will mean we lose the histories.  Thereore this
removes update_all and we now iterate through each allocation, set the
specific field to nil and re-save it; creating a new version.