Insert the data from the XLSX into the delius_data table
Previous PRs had implemented the XLSX parsing, and setting up of the
delius_data and tier_changes table.

This PR implements the writing of data into the delius_data table and
adds some useful helper functions to the DeliusData model.

During this work it was discovered that noms_no is not unique and so we
had to switch to matching on existing data based on the CRN.  This means
that we may end up with tier_changes which are not accurate, so we have
also added the CRN.

It is expected that we will add the CRN to the CaseInformation table so
that we can make this available in the offender models.