Create rake task to decrypt delius extract
As part of the continuous delius import we need to be able to take the
provided encrypted file and decrypt it with the given password, using
a msoffice-crypt command.  Some basic checks will be done to ensure an
input file, output file and password are present, without these the task
will fail.