Handle contact form submission
We have not yet decided on how we are going to deal with the
contact form data, so if the user submits a valid form we
just redirect to the contact page (GET). If the form is not
valid then we display the errors but the form will not be
submitted.