Remove prison from the case information code
Currently we use the prison code only to try and optimise the grouping
of offenders in the summary service, relying on it to fetch them all
once qnd then match them (or not) to the incoming offender list.

This is problematic as we are not the canonical source of an offender's
prison, and the optimisation was not necessarily faster than checking
the table for each batch of offenders being processed.

After this PR, the prison field still exists in the database (for now)
but is not used anywhere in the code.