Add recommended POM type column
Whilst working on displaying the override reasons in the allocation
history it was established that we need to know what the original
recommended POM type was at the time the primary POM was allocated. By
storing this data we can then add context around why an override took
place, and what POM type was originally recommended.  This PR just
adds the new column so work can start on updating the allocationVersion
to start recording this data.