Get the email addresses of primary poms in prisoner's allocation history
It is possible that a pom could have been allocated to a prisoner
more than once so we first get a unique list of the
primary_pom_nomis_ids in the allocation history to avoid calling the
API more than once.

We then call the API for each primary_pom_nomis_id in our list and grab
the first email we find (if present),  as it is possible that the poms
could have multiple email addresses.

We then return a hash the key being the primary_pom_nomis_id and the
value being the email address. If no email address is found for the pom
we return nil.