Display the primary pom email address in allocation history view
We now show the primary pom email address in the allocate and
reallocate partials for the allocation history views.

We no longer link to the primary pom page, we have a mail_to link
allowing the user to email the pom instead.

Now that we are displaying the primary pom email address in
allocation history view our test have to call the API to retrieve
the emails so we must have valid primary_pom_nomis_ids
otherwise the api will return a 404 status causing the tests to fail.

The allocation history feature test has been moved to a separate file.