Implement limit to reason text box for overrides
At present we have a free text field that SPOs can write in their
reasons for overriding a recommended POM type during an allocation.  A
conversation with our content designer and service designer led to an
agreement that we will restrict the number of characters that can be
added to no more than 175 before they see alerts.  If an SPO tries to
submit the form with the either/both text areas containing more than 175
characters then it will prevent them from progressing further and
include an error in the summary at the top of the page.