Re-instate the cache in tests so that they succeed
We had previously added `config.cache_store = :null_store` in an attempt
to remove caching in test, but it appears this introduced problems with
some of our longer running tests.  This is likely an issue with patterns
of usage and the T3 environment (e.g. it doesn't like large amounts of
traffic in short periods of time).

After this PR, tests run reliably.