Fix breakage when offender is allocated but not sentenced
Currently it is possible that an offender may be allocated, but not
sentenced. If this is the case, the responsibility service is unable to
work out if it is a new case or not.

When determining if the offender is a new case, it will now return true
if the offender is not sentenced.