Add a rake task for downloading the latest delius attachment
This PR adds a new rake task `delius_email:fetch_latest` that will
connect to the appropriate IMAP server, navigate to a specific folder
and then download the latest `application/zip attachment` it can find.

The single argument to the rake task, is the name to give to the
downloaded attachment.  The username, password and folder to check are
specified in environment variables.  These are `DELIUS_EMAIL_USERNAME`,
`DELIUS_EMAIL_PASSWORD`, and `DELIUS_EMAIL_FOLDER`.

To test that this works correctly the `Net::IMAP`` client is mocked and
the key methods implemented, working with json fixtures on disk so no
access to a working IMAP server is required.  The folders in the 'IMAP
server' are folders in the `fixtures/imap/` directory, each JSON file
represents a message where the first part of the filename represents the
IMAP message id.