Remove references to Allocations
At the start of this project we initially had a database table called
Allocations, however, we have since moved on to use AllcationVersions
instead.  Previous steps were taken to migrate the data and stop using
Allocations throughout the codebase, so this is the final step to remove
them completely, and drop the table.  Leaving it in will lead to
confusion.