Update AllocationVersion model re: releases
At present we are experiencing issues with our nightly Movement cron
job, which checks which offenders have been transferred or released.  We
have a validation on the AllocationVersion to check the presence of a
prison, but this will be nil if someone leaves prison.  Therefore, we
are now only validating the presence of prison IF the event trigger
matches 'offender_released', and are explicitly setting prison to nil in
this same case.