Adds a rake task to run the delius ETL
The included rake task `rake delius_etl:process` will run all of the
steps necessary to update the case information for offenders, from a
daily dump of delius data.

These steps mostly exist as existing rake tasks, and this task
integrates them so that they are all executed and any downloaded data is
cleaned up afterwards.

The steps it performs are:

* Download the latest attachment
* Unzip the attachment
* Decrypt the unzipped file
* Load the decrypted file into the database
* Create CaseInformation from the newly loaded data (pending).

As soon as the process has completed (or failed) the working folder
where files were stored is deleted.

The README.md file has been updated with required environment variables
for the ETL process.