Add content to Allocation History deallocation events
When a POM or a co-worker has been deallocated (this does
not include when a prisoner has been transferred), add
content detailing the name and email address (if present)
of the POM or co-worker.

Move the functionality of determining whether the POMs
email address is present or not from the views into a
helper method.