Fix Allocation History view
We have a small bug where if the "created_by_name" field is nil then the
allocation history page crashes.  There are a number of allocations that
were made before this field was implemented, so to ensure the page
doesn't crash we are only displaying the name if it exists.