Handle movements with no target agency
When we receive a movement that is a transfer, we are seeing some that
we believe have a `to-agency` of nil.  To handle this we will not
process these events as they are not valid movements.  The
movement_service will now ignore any movement events that do not have a
target and source agency.