Handle POMs without emails
When a POM has no email the send_email function crashes because it is
checking for what it believes is an empty list with `.empty?`. When this
is nil it crashes as `.empty?` is not a message nil understands.

Changed the call to `.blank?` which works with both nil and empty lists.