New API call to get all movements for a specific offender
As we need to clean up from the movement service not doing the correct
thing with some offenders, it is helpful to be able to get the entire
movement history (ADMissions, TRaNsfers and RELeases) for a single
offender.

The movements for this offender will be returned as `Movement` objects
in date order (oldest first).