When de-allocating an offender, do not reset the prison
Currently, when releasing an offender, we attempt to set the prison in
their allocation object to nil.

Rather than do this, we will leave their prison set to where they were
released from (but still, with no POMs) to act as a history.

If the offender is re-admitted to prison at a later stage, they will
have no case information (unless delius import has re-added it, in which
case it will be up-to-date). If the case information does exist, they
will show as ready to allocate (because they have no pom in their
allocation at this prison) but the history will show where they were
released from.