Don't process movements for remandees
Following an investigation into the MovementService we found that we
were processing movements for offenders on remand.  If they have
previously been in prison then we have old allocation records attributed
to them and this was causing issues where the offender is being
deallocated.  However, as they are on remand we shouldn't be processing
any movements.