Step 1: Prepare for Zendesk
We have reached a stage where we need to utilise Zendesk to manage
feedback from the users at prisons we've rolled out.  At present we are
getting direct emails to a group email address and it's becoming
increasingly difficult to track what the status is, whether or not
someone has responded and so forth.

We have been set up with a Zendesk account and want to ensure that we
don't lose crucial information so adding these in preparation for the
work around implementing Zendesk.