Speed up reallocation page (and get_poms).
Have re-instated the nplus one fix from Stephen, and also changed
add_detail so that it does less work.

Currently add_detail would fetch all the allocations for a POM, then
group them by allocated_at_tier, before then counting the size of the
list.  This would be slow (essentially fetching all the allocations for
a prison) and also memory intensive as we would be storing all those
allocations in RAM unnecessarily.

This PR changes the call to have the server group the response and just
return the counts.  This means the call should be a lot less variable
than previously and that we move a lot less data around.