Add search bar to summary pages
We have received feedback from users that they would like to have
better search functionality.  We have an establishment-wide search bar
in the main dashboard, but this can be laborious/easy to forget about if
users are spending a lot of time allocating/reallocation.  Therefore a
decision has been made to add the search bar to the allocated,
unallocated and pending summary pages to makes searching simpler, and
easier to access for staff.