Manually update rails-html-sanitizer gem
We have started to see deprecation warnings whilst running specs,
specifically "DEPRECATION WARNING: warning: white_list_sanitizer
isdeprecated, please use safe_list_sanitizer instead."  An issue was
opened on the Rails Github account
(https://github.com/rails/rails/issues/36858) and basically the fix is
to update to the latest version of the gem.

However, we have reduced the frequency we are getting Dependabot updates
and it could be another couple of weeks before this particular update
comes through, and therefore thought it would be a good idea to just
update it manually so we can stop seeing the deprecation warning.