[POM-298] Add debugging page
During the onboarding process we have received a number of queries from
users in regards to missing offenders, offenders with incorrect
information and so forth.  For example: for an offender to show up in
our service they need to be over 18, convicted, sentenced, etc.  At the
moment a dev needs to go and run a number of commands to try and find
out the various pieces of information and pass it along to whoever is on
duty.  Therefore, it was agreed that we should add a debugging page,
that anyone in the team can access, that provides a breakdown of the
individual offender to help us answer these types of queries.