Change the way get the allocated POM's name for an offender in SummaryService
We were using the method `set_allocated_pom_name` in OffenderService
but this is a painfully slow function call as it re-fetches the
allocation for each offender being shown in the allocated tab to
work out the name of the POM.

This commit removes the call to `set_allocated_pom_name` and grabs
the primary_pom_name and primary_pom_allocated_at date from the
allocations we have already obtained.

`OffenderService.set_allocated_pom_name` is used in SearchController
so it has been left untouched for now.