Add transfer check in MovementService
We are continuing to see a few cases failing the prison validation when
running the MovementService. The issue appears to be when an offender is
being brought into prison (movement_type: ADM & direction_code: IN), but
they were previously in prison and have a historical allocation. If it
is a new case then we want to bail out and not process any records.

Our initial assumption was that if the “from_agency” code was missing
then this would indicate it was an admission rather than a transfer,
and we wouldn’t process the admissions.  However, we are now seeing this
isn’t the case and we need to explicitly check if the “from_agency” is
from outside the prison estate.