[POM-281] Stop using the Custody API
We currently use the Custody API to retrieve some data that was
previously not available in Elite2.  However, whilst the API is going to
be supported, all new endpoints will be added to Elite2.  Therefore it
would be easier/simpler to use one API to retrieve the data we need.