Create SummaryService method `entered_prison_dates`
`entered_prison_dates` returns an array of hashes. The hash
contains the offender_no and the days_count.

The `days_count` value is generated from the
`MovementApi.movements_for method`, which returns information
on the most recent details of the prison a given offender is
located at. This information includes the `create_date_time`
which is the date that the offender was allocated to that prison.

The determination of the `days_count` value depends whether the
offender arrived at the prison before or after the
`omic_start_date` (4th Feb 2019). If the date is before this,
then the `omic_start_date` is set as the first day
otherwise it is the `create_date_time`.