Show POM names in the correct format (regardless of old data)
Previously we were storing the pom names in the allocation as
"first_name last_name" however when we display it on the page, we show
it as "last_name, first_name".

This PR change how we store the name in the allocation, and also adds a
helper to ensure we can handle the previously set data.