Establish if offender has been recalled
We need to know whether an offender has been recalled as if this is the
case then how long they have left to serve will be used to determine
whether case responsibility remains in the community, or is transferred
back to custody.  At present, we don't have the logic for this so have
added the recall check to the debugging page as a starting point.

Another ticket will be added to the backlog so the team can understand
the exact logic required around the responsibility calculation.