Refactor how the API responses are deserialised
Currently we are using active record attributes to defined what the
models contain, and using respond_to and public_send to set the values
on the model.  Unfortunately this means we spend a lot of time checking
if the model responds to fields that we are not interested in, and each
field then contains a lot of unnecessary data (the old value, the
typecase value etc).

This PR changes how we deserialise the data to make each model
responsible for only pulling the fields that it needs.  Each field then
becomes a standard ruby attribute and therefore uses less RAM.

WARNING: The cache will need to be cleared after this PR has been
deployed as we currently cache the PageMeta object rather than as JSON.