A simple rake task to fix up incorrect allocations
As we currently have allocations that are incorrect, because of releases
and transfers that the movement_service has missed, we need to clean
them up.

This version of a rake task will, for a single prison at a time, fix up
all allocations for that prison to ensure offenders no longer there are
deallocated (and then should be available for allocation in their new
prison, if they have one).

Initially we will simply print out what we would have done as we don't
want to risk breaking a production database.