Fix bad allocations
This PR builds on the previous fix-all-the-data PR to actually change
the records that we identify as bad, by deallocating offenders.

It continues to write to stdout so we know what happened.

As a side-effect (the offender can't be found) we also identify nomis
IDs that have been merged into another ID although we don't do anything
about that at present.  We need to discuss whether it is safe to discard
these allocations if NOMIS doesn't know about the nomis id.