Home breadcrumb to dashboard not root
Currently the home link in the dashboard sends the user to the root,
which means if they have a different active_caseload (in NOMIS) then
they can end up looking at a different prison from the one they have
just been working in.

This PR changes home to point to the dashboard for the current prison.