Fix allocation date exceptions
Allocation dates are occasionally raising exceptions. This is due the
`primary_pom_allocated_at` date being nil and calling the method
`.to_date` on it.

This PR fixes this issue by checking whether `primary_pom_allocated_at`
date for an allocated offender is nil, if so we use the `updated_at`
date instead and only call. The `.to_date` method is only called if
the date is not nil.