Handover date calculation
Currently the calculation of the offender’s handover date is done in a view helper.
This PR moves the handover calculations to a service - HandoverDateService. The tests
associated with the handover calculations have been moved out of the helper_spec and into
the service_spec.

The handover methods have now been added to the OffenderBase class and called directly
in the views.