SPIKE: Refactor how we do services
The current approach to services results in us having a single class
containing many different functions (typically on a singleton - eek) and
this makes it hard to find a specific function to change, and hard to
test.

This PR contains a refactor of how we do services, specifically it
replaces PrisonOffenderManagerService with a module called POM that
contains one callable class per function.  These then map to one spec
file per function.  e.g. app/services/pom/get_pom maps to
spec/services/pom/get_pom_spec.

It is thought that this approach is a lot cleaner, at the cost of trying
to find nice names for things - but at least it should be reasonably
simple to rename the new callable classes.